requirements.txt
# Requirements - PDF RAG Chatbot

# Core framework
streamlit==1.28.1
langchain==0.0.350
langchain-community==0.0.1

# PDF processing
PyMuPDF==1.23.14
pdfplumber==0.10.3
python-magic==0.4.27

# Vector search and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# OpenAI integration
openai==1.3.8
tiktoken==0.5.2

# Text processing
nltk==3.8.1
spacy==3.7.2

# Data handling
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Logging and monitoring
loguru==0.7.2

# Development tools
black==23.12.1
flake8==6.1.0
pytest==7.4.3
pytest-cov==4.1.0

# Optional: Local LLM support
# transformers==4.36.2
# torch==2.1.2

# Optional: Azure Document Intelligence
# azure-ai-formrecognizer==3.3.0
# azure-identity==1.15.0